create colorpallete
