+++ demo/src/components/ChannelOptions.js
@@ -6,75 +6,33 @@ import Paper from '@material-ui/core/Paper';
 import Popper from '@material-ui/core/Popper';
 import MenuItem from '@material-ui/core/MenuItem';
 import MenuList from '@material-ui/core/MenuList';
-import LensIcon from '@material-ui/icons/Lens';
 
 import { makeStyles } from '@material-ui/core/styles';
 
-import { COLOR_PALLETE } from '../constants';
+import ColorPalette from './ColorPalette';
 
 const useStyles = makeStyles(theme => ({
   paper: {
     backgroundColor: 'rgba(0, 0, 0, 0.75)'
   },
-  colorRow: {
+  span: {
     width: '70px',
-    height: '40px',
-    display: 'flex',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    flexWrap: 'wrap'
+    textAlign: 'center',
+    paddingLeft: '2px',
+    paddingRight: '2px'
   },
-  colorBlock: {
-    padding: '3px',
-    width: '16px',
-    height: '16px'
-  },
-  menu: {
-    width: '70px',
-    textAlign: 'center'
-  },
-  button: {
-    width: '17px',
-    height: '17px'
-  },
-  colorMenuItem: {
+  colors: {
     '&:hover': {
       backgroundColor: 'transparent'
     },
     paddingLeft: '2px',
     paddingRight: '2px'
-  },
-  menuItem: {
-    paddingLeft: '2px',
-    paddingRight: '2px'
   }
 }));
-
-const ColorPalette = ({ handleChange }) => {
-  const classes = useStyles();
-  return (
-    <div className={classes.colorRow} aria-label="color-swatch">
-      {COLOR_PALLETE.map(color => {
-        return (
-          <IconButton className={classes.colorBlock}>
-            <LensIcon
-              fontSize="small"
-              style={{ color: `rgb(${color})` }}
-              key={color}
-              onClick={() => handleChange(color)}
-              className={classes.button}
-            />
-          </IconButton>
-        );
-      })}
-    </div>
-  );
-};
 
 function ChannelOptions({ handleChange }) {
   const [open, toggle] = useReducer(v => !v, false);
   const anchorRef = useRef(null);
-  const classes = useStyles();
 
   const handleColorSelect = color => {
     handleChange('CHANGE_COLOR', color);
@@ -85,6 +43,7 @@ function ChannelOptions({ handleChange }) {
     handleChange('REMOVE_CHANNEL');
   };
 
+  const classes = useStyles();
   return (
     <>
       <IconButton
@@ -95,19 +54,14 @@ function ChannelOptions({ handleChange }) {
       >
         <MoreVertIcon fontSize="small" />
       </IconButton>
-      <Popper
-        open={open}
-        role={undefined}
-        anchorEl={anchorRef.current}
-        placement="bottom-end"
-      >
+      <Popper open={open} anchorEl={anchorRef.current} placement="bottom-end">
         <Paper className={classes.paper}>
           <ClickAwayListener onClickAway={toggle}>
             <MenuList id="channel-options">
               <MenuItem dense disableGutters onClick={handleRemove}>
-                <span className={classes.menu}>Remove</span>
+                <span className={classes.span}>Remove</span>
               </MenuItem>
-              <MenuItem dense disableGutters className={classes.colorMenuItem}>
+              <MenuItem dense disableGutters className={classes.colors}>
                 <ColorPalette handleChange={handleColorSelect} />
               </MenuItem>
             </MenuList>
